Updated the `input_module_opencti_stream.py` to set the `_time` field to the current time when ingesting data. This ensures that the events reflect the time they were processed rather than the time they were created in OpenCTI.
update the `props.conf` to set the `TIME_PREFIX`
to `^` for the `opencti_stream` sourcetype.
This change ensures that the `_time` field is
correctly set to the current time when data is
ingested, rather than using the time from the
OpenCTI data.
This is necessary to ensure that the events
reflect the time they were processed rather than
the time they were created in OpenCTI.
This change is necessary to ensure that the
events reflect the time they were processed
rather than the time they were created in OpenCTI.

Updated savedsearches.conf to ensure that the
searches are correctly configured to use the
new `_time` field setting. This ensures that
the searches return results based on the time
the events were processed, rather than the time
they were created in OpenCTI.
This change is necessary to ensure that the
searches return results based on the time the
events were processed, rather than the time
they were created in OpenCTI.